src/senzey_bots/pyproject.toml
[project]
name = "senzey-bots"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "streamlit==1.54.0",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "python-dotenv>=1.0.0",
    "pydantic==2.12.5",
    "sqlalchemy==2.0.47",
    "alembic==1.18.4",
    "argon2-cffi==25.1.0",
    "cryptography>=43.0.0",
    "toml>=0.10.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/senzey_bots"]

[dependency-groups]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "responses>=0.25.0",
    "pytest-cov>=7.0.0",
]
